show non trainable params in summary, use pandas to display engineering units
This shows a trainable parameter in the full data frame and trainable vs non-trainable params in the summary. This should be like keras [example](https://github.com/keras-team/keras/pull/11795#issue-235955201).

I also changed the pandas display so that instead of putting (M) or (K) manually in the label, pandas appends it to the number automatically.